Who am I?
Hi there, 
I am a student/ undergraduate at the American college Kandy campus.  I'm an enthusiast in Artificial Intelligence as well as in web development. python is my main interest when it's up to coding.  I am 19 years old and I am a rugger player. My main hobby since I was a kid was collecting collectibles, and next, it's coding. Coding is ones and zeros, It's a secret message we built up to interact with the computer. In my words, It's a secret message we share with the computer. It's real and it's alive. Coding is life.